MySQL driver cleanup

+++ driver/mysql/mysql.go
@@ -13,6 +13,8 @@ import (
 	"github.com/rubiojr/kv/types"
 )
 
+const insert = "?"
+
 type Database struct {
 	t  string
 	db *sql.DB
@@ -101,14 +103,14 @@ func (d *Database) Set(key string, value []byte, expiresAt *time.Time) error {
 }
 
 func (d *Database) MGet(keys ...string) ([][]byte, error) {
-	t := time.Now().UTC()
-	args := make([]interface{}, len(keys))
-	for i, id := range keys {
-		args[i] = id
+	if len(keys) < 1 {
+		return [][]byte{}, nil
 	}
-	sql := fmt.Sprintf("SELECT `key`, value FROM %s WHERE `key` IN(?"+strings.Repeat(",?", len(args)-1)+") AND (`expires_at` IS NULL OR `expires_at` > '%s')", d.t, t)
 
-	rows, err := d.db.Query(sql, args...)
+	knames, inserts := vRow(keys...)
+	sql := fmt.Sprintf("SELECT `key`, value FROM %s WHERE `key` IN(%s) AND (`expires_at` IS NULL OR `expires_at` > '%s')", d.t, inserts, time.Now().UTC())
+
+	rows, err := d.db.Query(sql, knames...)
 	if err != nil {
 		return nil, err
 	}
@@ -118,7 +120,7 @@ func (d *Database) MGet(keys ...string) ([][]byte, error) {
 	for rows.Next() {
 		var key, value string
 		if err := rows.Scan(&key, &value); err != nil {
-			panic(err)
+			return [][]byte{}, err
 		}
 		values = append(values, []byte(value))
 	}
@@ -127,20 +129,15 @@ func (d *Database) MGet(keys ...string) ([][]byte, error) {
 }
 
 func (d *Database) MDel(keys ...string) error {
-	const insert = "?"
-	var klist []string
-	values := []interface{}{}
-
-	for _, k := range keys {
-		klist = append(klist, insert)
-		values = append(values, k)
+	if len(keys) < 1 {
+		return nil
 	}
 
-	inserts := strings.Join(klist, ",")
+	knames, inserts := vRow(keys...)
 
 	sql := fmt.Sprintf("DELETE FROM %s WHERE `key` IN(%s)", d.t, inserts)
 
-	_, err := d.db.Exec(sql, values...)
+	_, err := d.db.Exec(sql, knames...)
 	return err
 }
 
@@ -156,17 +153,15 @@ func (d *Database) MExists(keys ...string) ([]bool, error) {
 		return []bool{}, nil
 	}
 
-	now := time.Now().UTC()
-	args := make([]interface{}, lkeys)
+	knames, inserts := vRow(keys...)
 	mcheck := map[string]bool{}
-	for i, id := range keys {
-		args[i] = id
+	for _, id := range keys {
 		mcheck[id] = false
 	}
 
-	sql := fmt.Sprintf("SELECT `key` FROM %s WHERE `key` IN(?"+strings.Repeat(",?", lkeys-1)+") AND (`expires_at` IS NULL OR `expires_at` > '%s')", d.t, now)
+	sql := fmt.Sprintf("SELECT `key` FROM %s WHERE `key` IN(%s) AND (`expires_at` IS NULL OR `expires_at` > '%s')", d.t, inserts, time.Now().UTC())
 
-	rows, err := d.db.Query(sql, args...)
+	rows, err := d.db.Query(sql, knames...)
 	if err != nil {
 		return nil, err
 	}
@@ -196,4 +191,15 @@ func (d *Database) Exists(key string) (bool, error) {
 		return false, err
 	}
 	return err == nil && values[0], err
+}
+
+func vRow(keys ...string) ([]interface{}, string) {
+	var ilist []string
+	knames := []interface{}{}
+	for _, k := range keys {
+		ilist = append(ilist, insert)
+		knames = append(knames, k)
+	}
+
+	return knames, strings.Join(ilist, ",")
 }
